# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AcceptContactInvitationDto": {
        "properties": {
          "consent": {
            "$ref": "#/components/schemas/ContactConsentDto"
          },
          "password": {
            "maxLength": 512,
            "minLength": 12,
            "type": "string",
            "writeOnly": true
          },
          "privateKeyEnvelope": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ContactPrivateKeyEnvelopeDto"
              }
            ],
            "writeOnly": true
          },
          "token": {
            "minLength": 40,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "token",
          "password",
          "privateKeyEnvelope",
          "consent"
        ],
        "type": "object"
      },
...
      "AuditDetailDto": {
        "properties": {
          "password": {
            "minLength": 1,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "password"
        ],
        "type": "object"
      },
...
      "ChangeContactPasswordDto": {
        "properties": {
          "newPassword": {
            "maxLength": 512,
            "minLength": 12,
            "type": "string",
            "writeOnly": true
          },
          "newPrivateKeyEnvelope": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ContactPrivateKeyEnvelopeDto"
              }
            ],
            "writeOnly": true
          },
          "oldPassword": {
            "maxLength": 512,
            "minLength": 12,
            "type": "string",
            "writeOnly": true
          },
          "token": {
            "minLength": 40,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "oldPassword",
          "newPassword",
          "newPrivateKeyEnvelope"
        ],
        "type": "object"
      },
...
      "ContactConsentDto": {
        "properties": {
          "denialDisclosureAccepted": {
            "type": "boolean"
          },
          "documentSha256": {
            "pattern": "^[0-9a-f]{64}$",
            "type": "string"
          },
          "privacyAccepted": {
            "type": "boolean"
          },
          "stage2LockAccepted": {
            "type": "boolean"
          },
          "termsAccepted": {
            "type": "boolean"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "version",
          "documentSha256",
          "termsAccepted",
          "privacyAccepted",
          "denialDisclosureAccepted",
          "stage2LockAccepted"
        ],
        "type": "object"
      },
...
      "ContactLoginDto": {
        "properties": {
          "displayName": {
            "minLength": 1,
            "type": "string"
          },
          "entryToken": {
            "type": "string",
            "writeOnly": true
          },
          "password": {
            "maxLength": 512,
            "minLength": 12,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "displayName",
          "password"
        ],
        "type": "object"
      },
      "ContactPrivateKeyEnvelopeDto": {
        "properties": {
          "ciphertext": {
            "format": "byte",
            "type": "string",
            "writeOnly": true
          },
          "kdfParams": {
            "type": "object",
            "writeOnly": true
          },
          "kdfSalt": {
            "format": "byte",
            "type": "string",
            "writeOnly": true
          },
          "nonce": {
            "format": "byte",
            "type": "string",
            "writeOnly": true
          },
          "privateKeyProof": {
            "format": "byte",
            "type": "string",
            "writeOnly": true
          },
          "publicKey": {
            "format": "byte",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "publicKey",
          "ciphertext",
          "nonce",
          "kdfSalt",
          "kdfParams",
          "privateKeyProof"
        ],
        "type": "object"
      },
...
      "CreateContactInvitationDto": {
        "properties": {
          "displayName": {
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "format": "email",
            "type": "string"
          }
        },
        "required": [
          "displayName",
          "email"
        ],
        "type": "object"
      },
...
      "RemoveContactDto": {
        "properties": {
          "password": {
            "minLength": 1,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "password"
        ],
        "type": "object"
      },
      "RequestContactPasswordChangeDto": {
        "properties": {
          "password": {
            "minLength": 1,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "password"
        ],
        "type": "object"
      },
      "ResolveContactInvitationDto": {
        "properties": {
          "token": {
            "minLength": 40,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "token"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Local V1 HTTP contract",
    "title": "Digital Legacy System API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/auth/contact/login": {
      "post": {
        "operationId": "ContactAuthController_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactLoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Contact session created"
          }
        },
        "summary": "Login an emergency contact",
        "tags": [
          "Contact authentication"
        ]
      }
    },
...
    "/contact-invitations/accept": {
      "post": {
        "operationId": "ContactInvitationsController_accept",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptContactInvitationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Accept an invitation with consent and a wrapped private key",
        "tags": [
          "Contact invitations"
        ]
      }
    },
    "/contact-invitations/resolve": {
      "post": {
        "operationId": "ContactInvitationsController_resolve",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveContactInvitationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Resolve an invitation token supplied in the request body",
        "tags": [
          "Contact invitations"
        ]
      }
    },
...
    "/contacts/password-change/complete": {
      "post": {
        "operationId": "ContactPasswordController_complete",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeContactPasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Rotate the contact password and wrapped private key",
        "tags": [
          "Contact password"
        ]
      }
    },
...
    "/owner/audit-events": {
      "get": {
        "operationId": "AuditController_list",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "eventType",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "result",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Read a redacted page of immutable private audit events",
        "tags": [
          "Owner private audit"
        ]
      }
    },
    "/owner/audit-events/{eventId}/detail": {
      "post": {
        "operationId": "AuditController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditDetailDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Reauthenticate before reading digest-only audit detail",
        "tags": [
          "Owner private audit"
        ]
      }
    },
...
    "/owner/contacts/{contactId}/invitation/resend": {
      "post": {
        "operationId": "ContactInvitationsController_resend",
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Revoke and resend a contact invitation",
        "tags": [
          "Contact invitations"
        ]
      }
    },
    "/owner/contacts/{contactId}/password-change-invitation": {
      "post": {
        "operationId": "ContactInvitationsController_requestPasswordChange",
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestContactPasswordChangeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Issue a one-time contact password-change invitation",
        "tags": [
          "Contact invitations"
        ]
      }
    },
    "/owner/contacts/{contactId}/remove": {
      "post": {
        "operationId": "ContactInvitationsController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveContactDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Remove an emergency contact and require share regeneration",
        "tags": [
          "Contact invitations"
        ]
      }
    },
    "/owner/contacts/invitations": {
      "post": {
        "operationId": "ContactInvitationsController_invite",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactInvitationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Invitation queued"
          }
        },
        "summary": "Create a one-time emergency contact invitation",
        "tags": [
          "Contact invitations"
        ]
      }
    },
...
  },
  "servers": [],
...
}